Dirichlet-sample of a multinomial election poll

Calculate a dirichlet-sample of a multinomial election poll

sample_dirichlet(vote = c(cdu = 0.5, spd = 0.4, fdp = 0.1), sample_n = 1000,
  n_draw = 10000)

Arguments

vote
A labeled party vote share vector.

sample_n
The number of observations in the poll sample.

n_draw
How many samples to draw from the dirichlet distribution.

Value

A data frame containing n rows of samples for each party.

Warning

Do not operate heavy machinery within 8 hours of using this function.

Examples

sample_dirichlet(vote = c(cdu = 0.33,
                          spd = 0.20,
                          fdp = 0.11,
                          linke = 0.09,
                          gruene = 0.09,
                          afd = 0.12,
                          sonstige = 0.05),
                 sample_n = 1000,
                 n_draw = 10000)

#>             cdu       spd        fdp      linke     gruene        afd
#> 1     0.3263399 0.1988020 0.11726191 0.10361851 0.08665359 0.12515845
#> 2     0.3337058 0.1959155 0.09722950 0.09839372 0.08617490 0.12420336
#> 3     0.3215732 0.2086351 0.09814464 0.11535276 0.08854444 0.11161556
#> 4     0.3357623 0.1919415 0.11002978 0.09072780 0.10426308 0.11929155
#> 5     0.3285782 0.1954696 0.10665991 0.08371518 0.10348364 0.11117427
#> 6     0.3370559 0.2119634 0.11298991 0.07379723 0.10014057 0.11340973
#> 7     0.3351886 0.2105782 0.08835404 0.09090169 0.09829719 0.12519659
#> 8     0.3530302 0.1887842 0.09688835 0.09291941 0.10307218 0.12084120
#> 9     0.3115666 0.1914404 0.11502663 0.10518208 0.09708737 0.12964639
#> 10    0.3117649 0.2234992 0.12127957 0.09497880 0.09312821 0.11153237
#> 11    0.3272175 0.1965489 0.11417137 0.10000428 0.09986031 0.11927982
#> 12    0.3358522 0.1890380 0.11061838 0.10963905 0.07572157 0.11534151
#> 13    0.3290292 0.2068582 0.12726557 0.09216065 0.09227155 0.10327558
#> 14    0.3227993 0.2167637 0.10192069 0.07456855 0.08762935 0.14339755
#> 15    0.3511326 0.2031886 0.10934687 0.08867160 0.09419730 0.11529453
#> 16    0.3177470 0.2154254 0.11386028 0.08928855 0.09649951 0.11344913
#> 17    0.3148194 0.2046448 0.09427036 0.10055133 0.10274419 0.13945310
#> 18    0.3311613 0.2258122 0.11613283 0.09232527 0.08051082 0.09867530
#> 19    0.3502955 0.2063354 0.09635911 0.07683513 0.11403681 0.11613974
#> 20    0.3004486 0.2107896 0.10971431 0.10678447 0.09455801 0.12497785
#> 21    0.3413100 0.2002611 0.10694937 0.09983834 0.07651988 0.11082140
#> 22    0.3076044 0.2046151 0.12958302 0.08686377 0.08384889 0.13053844
#> 23    0.3168091 0.2189246 0.12009822 0.06504710 0.09150188 0.13000531
#> 24    0.3528267 0.2003654 0.10461899 0.09003729 0.08930107 0.11162851
#> 25    0.3778105 0.1828024 0.10830316 0.07757031 0.08919309 0.11206604
#> 26    0.3195957 0.2189809 0.11589260 0.08179115 0.09543702 0.11454440
#> 27    0.3341729 0.2203685 0.09890167 0.08887114 0.10165755 0.10830404
#> 28    0.3288942 0.1998238 0.09684994 0.10819947 0.08525857 0.13918194
#> 29    0.3329414 0.1947274 0.11740721 0.08335327 0.07841681 0.14179529
#> 30    0.3398759 0.2063184 0.11850547 0.09205722 0.07919140 0.11453174
#> 31    0.3282107 0.2181101 0.11569964 0.07885455 0.09616710 0.12590443
#> 32    0.3425076 0.1807721 0.12087122 0.10326279 0.08955391 0.11901285
#> 33    0.3249342 0.2058912 0.11200535 0.09226511 0.09971096 0.10856186
#> 34    0.3604811 0.1785224 0.11438169 0.07895764 0.11238629 0.11045514
#> 35    0.3313731 0.2239924 0.10284020 0.09136683 0.09632232 0.10869153
#> 36    0.3540640 0.1855390 0.09492445 0.09356642 0.09405639 0.12449171
#> 37    0.3213696 0.2125819 0.10488987 0.08673897 0.09402278 0.12105229
#> 38    0.3254554 0.2269325 0.09306670 0.08214350 0.09143682 0.12471105
#> 39    0.3087778 0.2129086 0.13074314 0.09222616 0.08290230 0.12558898
#> 40    0.3330295 0.1988920 0.09768901 0.09797163 0.09591238 0.12343004
#> 41    0.3493039 0.1830308 0.10717365 0.07540862 0.10295164 0.13564750
#> 42    0.3296130 0.1931488 0.10575687 0.10031211 0.11063390 0.12194644
#> 43    0.3353832 0.1917830 0.11874837 0.08850704 0.09402643 0.12349104
#> 44    0.3325874 0.2148001 0.10022743 0.08935781 0.08336682 0.12822278
#> 45    0.3309500 0.2137117 0.10285971 0.10601775 0.09296859 0.10864903
#> 46    0.3256108 0.1856820 0.12875803 0.09170740 0.10943769 0.11329577
#> 47    0.3299612 0.1847060 0.14443616 0.09127269 0.07068651 0.12247450
#> 48    0.3523777 0.1792878 0.11375148 0.07293079 0.09014510 0.13132587
#> 49    0.3056351 0.2305785 0.11319830 0.09215089 0.08423911 0.12371591
#> 50    0.3176477 0.2030315 0.10501934 0.08585264 0.09527515 0.13095701
#> 51    0.3488974 0.2044353 0.10049588 0.09459875 0.09777013 0.10970671
#> 52    0.3375812 0.1915625 0.11235933 0.08014677 0.08962918 0.13224398
#> 53    0.3223568 0.2230927 0.11070729 0.10449623 0.07592212 0.11482805
#> 54    0.3238500 0.1854582 0.11814698 0.08496344 0.09575179 0.13019653
#> 55    0.3067614 0.1975146 0.12286571 0.09997577 0.09884463 0.11736573
#> 56    0.3473395 0.2219617 0.10311350 0.06823792 0.10876615 0.09842763
#> 57    0.3241816 0.1866574 0.11836896 0.09943663 0.08973414 0.12009568
#> 58    0.3337810 0.2088392 0.11443391 0.08048215 0.09344271 0.12383773
#> 59    0.3536942 0.1839955 0.12170555 0.08604862 0.08840693 0.11564512
#> 60    0.3123491 0.2082746 0.12656735 0.08289222 0.09641798 0.12775357
#> 61    0.3345983 0.1888672 0.09987280 0.09178296 0.09673252 0.13090845
#> 62    0.3194375 0.2143191 0.10612816 0.10491666 0.08963479 0.12203297
#> 63    0.3270125 0.2036429 0.11661458 0.09098869 0.10472516 0.11380557
#> 64    0.3504915 0.1925213 0.10081515 0.07707759 0.10527465 0.11210483
#> 65    0.3312445 0.2211319 0.11285445 0.09380417 0.08207485 0.11249491
#> 66    0.3313598 0.1970152 0.10850407 0.10413558 0.09291183 0.11584350
#> 67    0.3106996 0.2155088 0.10242045 0.09014560 0.10056423 0.12429513
#> 68    0.3210076 0.2157032 0.11210116 0.08943194 0.08944665 0.12499057
#> 69    0.3352896 0.1805734 0.09485909 0.09884608 0.09298568 0.14915534
#> 70    0.3258590 0.2192278 0.10769136 0.09784191 0.08668737 0.11028175
#> 71    0.3362022 0.1785367 0.11975721 0.07863932 0.09804812 0.13382674
#> 72    0.3290116 0.1964948 0.11586279 0.09041188 0.08941912 0.11478949
#> 73    0.3287826 0.1881444 0.11572634 0.08630229 0.10004634 0.13148354
#> 74    0.3492027 0.1899960 0.10859303 0.09077077 0.08193915 0.12770869
#> 75    0.3571412 0.1962580 0.10901047 0.09124515 0.08579137 0.10444557
#> 76    0.3241602 0.2128604 0.10101312 0.09299918 0.08603069 0.12767807
#> 77    0.3760049 0.1698676 0.10853074 0.08347037 0.09037136 0.11384245
#> 78    0.3305828 0.2104709 0.09300205 0.08387190 0.09580664 0.11797266
#> 79    0.3403209 0.2164245 0.11091603 0.07864912 0.09550012 0.11850762
#> 80    0.3440539 0.1785445 0.11897924 0.08592506 0.11590860 0.11150764
#> 81    0.3172846 0.2190542 0.11559558 0.11363555 0.09187586 0.09982418
#> 82    0.3575610 0.1811282 0.11360633 0.08555045 0.09965742 0.10547098
#> 83    0.3494831 0.2114701 0.10097038 0.09537849 0.09060516 0.10672476
#> 84    0.3350537 0.1906570 0.11673560 0.10019136 0.09139718 0.10889064
#> 85    0.3322495 0.2016323 0.11146481 0.09026043 0.08557599 0.12723721
#> 86    0.3425896 0.1934622 0.10430603 0.09939401 0.09425624 0.12527266
#> 87    0.3224066 0.2034805 0.12392179 0.09253270 0.07487394 0.12534457
#> 88    0.3262784 0.2115547 0.10897433 0.09550086 0.09149345 0.12731072
#> 89    0.3588607 0.1927075 0.10247967 0.10198658 0.08746038 0.10965224
#> 90    0.3135703 0.2138930 0.08963239 0.10351269 0.08801014 0.13559631
#> 91    0.3281347 0.2057313 0.10640127 0.10440531 0.08527201 0.12676952
#> 92    0.3292920 0.1957980 0.11267438 0.08289717 0.08907280 0.13552587
#> 93    0.3223228 0.2245931 0.12594086 0.08137823 0.08368257 0.10197659
#> 94    0.3364613 0.1870517 0.12470188 0.10295063 0.08617788 0.10417265
#> 95    0.3577464 0.1805878 0.09789241 0.10208537 0.10016252 0.11219298
#> 96    0.3156077 0.1789324 0.10617965 0.08581430 0.12160097 0.14122820
#> 97    0.3441688 0.2105724 0.08632983 0.09328948 0.09500177 0.11663414
#> 98    0.3431401 0.1945781 0.10543343 0.09931085 0.09420247 0.11385689
#> 99    0.3255337 0.1926467 0.12095619 0.09069705 0.09859886 0.12829849
#> 100   0.3278890 0.1993187 0.10990306 0.08548685 0.10532249 0.12020436
#> 101   0.3400361 0.1771533 0.12166747 0.09095951 0.09070509 0.13578025
#> 102   0.3605960 0.2021200 0.10392954 0.07927598 0.08749005 0.11627664
#> 103   0.3391937 0.2036155 0.10153755 0.09413397 0.08117095 0.12521725
#> 104   0.3493650 0.2033717 0.10759112 0.07877348 0.09982456 0.11512878
#> 105   0.3375173 0.2071162 0.12863682 0.09191266 0.08007535 0.11318715
#> 106   0.3559890 0.2134160 0.10362015 0.06441617 0.10251855 0.12169966
#> 107   0.3481442 0.1930012 0.10091297 0.09657594 0.08700696 0.12154147
#> 108   0.3475020 0.1993322 0.11741802 0.08473417 0.09268961 0.09700097
#> 109   0.3291347 0.2042185 0.12875635 0.08741503 0.08947610 0.11933718
#> 110   0.3240663 0.2053776 0.10242710 0.08860097 0.09169380 0.13498459
#> 111   0.3161064 0.2097043 0.13854128 0.09780747 0.07941592 0.10114784
#> 112   0.3508457 0.2089434 0.10996128 0.09289730 0.08180792 0.10429402
#> 113   0.3486926 0.1867723 0.09639841 0.09981278 0.08478642 0.12338331
#> 114   0.3398458 0.2129415 0.10932968 0.09259651 0.09268536 0.11151057
#> 115   0.3206637 0.2183579 0.11111021 0.08213262 0.08360684 0.12053181
#> 116   0.3591201 0.1964184 0.11403293 0.08839266 0.08582392 0.11727064
#> 117   0.3287632 0.1956186 0.12389301 0.10863739 0.07784946 0.11120253
#> 118   0.3316098 0.1887305 0.12629318 0.09683613 0.09155811 0.12269531
#> 119   0.3294250 0.1984438 0.11092260 0.10365011 0.09371378 0.11729947
#> 120   0.3423305 0.1931888 0.11626349 0.08964615 0.09747374 0.11602470
#> 121   0.3486551 0.1914901 0.11312418 0.08269261 0.10412055 0.11412331
#> 122   0.3472789 0.1827668 0.10442833 0.10580952 0.10592923 0.10483893
#> 123   0.3636060 0.2054285 0.10480278 0.08584541 0.07898484 0.11885895
#> 124   0.3585092 0.2082030 0.10328324 0.09086309 0.07322166 0.11073228
#> 125   0.3525297 0.1858755 0.12532290 0.10070000 0.08022721 0.11225101
#> 126   0.3355845 0.2094980 0.12650311 0.08702789 0.08416592 0.11579632
#> 127   0.3532063 0.2160855 0.11035051 0.08446041 0.08261287 0.10599319
#> 128   0.3332022 0.1930946 0.10870318 0.07953623 0.10145359 0.13387039
#> 129   0.3122128 0.2077861 0.09931186 0.10005402 0.09494860 0.12155635
#> 130   0.3314889 0.2095668 0.10667699 0.08372868 0.09353718 0.12310003
#> 131   0.3186087 0.1963124 0.10733556 0.08702479 0.10235416 0.13030338
#> 132   0.3177418 0.2142934 0.11691259 0.07810081 0.09647511 0.12783411
#> 133   0.3165561 0.2185711 0.11712537 0.08352067 0.08751587 0.11938770
#> 134   0.3417258 0.1826350 0.11994885 0.08550684 0.09848772 0.12011879
#> 135   0.3176006 0.2053788 0.10850208 0.09221751 0.08274753 0.13198515
#> 136   0.3253337 0.1915418 0.13168864 0.07750016 0.08983012 0.12804909
#> 137   0.3056943 0.1957284 0.12455043 0.10941247 0.09073570 0.12201073
#> 138   0.3443747 0.1932520 0.10963484 0.08744770 0.09364986 0.11896785
#> 139   0.3307163 0.1888285 0.10992812 0.09145257 0.10896180 0.10900637
#> 140   0.3290325 0.2093957 0.11227572 0.07694682 0.09602973 0.13167610
#> 141   0.3251267 0.2200313 0.10650564 0.09386276 0.08152643 0.13188514
#> 142   0.3616470 0.1969685 0.10960966 0.08808547 0.07952529 0.11557004
#>         sonstige
#> 1     0.04216572
#> 2     0.06437730
#> 3     0.05613430
#> 4     0.04798398
#> 5     0.07091924
#> 6     0.05064331
#> 7     0.05148371
#> 8     0.04446449
#> 9     0.05005049
#> 10    0.04381698
#> 11    0.04291783
#> 12    0.06378926
#> 13    0.04913922
#> 14    0.05292091
#> 15    0.03816849
#> 16    0.05373009
#> 17    0.04351679
#> 18    0.05538234
#> 19    0.03999831
#> 20    0.05272715
#> 21    0.06429991
#> 22    0.05694637
#> 23    0.05761370
#> 24    0.05122201
#> 25    0.05225452
#> 26    0.05375822
#> 27    0.04772417
#> 28    0.04179208
#> 29    0.05135872
#> 30    0.04951989
#> 31    0.03705347
#> 32    0.04401957
#> 33    0.05663130
#> 34    0.04481575
#> 35    0.04541368
#> 36    0.05335798
#> 37    0.05934457
#> 38    0.05625402
#> 39    0.04685296
#> 40    0.05307545
#> 41    0.04648388
#> 42    0.03858885
#> 43    0.04806091
#> 44    0.05143765
#> 45    0.04484316
#> 46    0.04550830
#> 47    0.05646299
#> 48    0.06018127
#> 49    0.05048216
#> 50    0.06221667
#> 51    0.04409588
#> 52    0.05647699
#> 53    0.04859682
#> 54    0.06163311
#> 55    0.05667216
#> 56    0.05215357
#> 57    0.06152563
#> 58    0.04518328
#> 59    0.05050409
#> 60    0.04574516
#> 61    0.05723778
#> 62    0.04353091
#> 63    0.04321065
#> 64    0.06171502
#> 65    0.04639513
#> 66    0.05023001
#> 67    0.05636618
#> 68    0.04731892
#> 69    0.04829081
#> 70    0.05241085
#> 71    0.05498968
#> 72    0.06401025
#> 73    0.04951451
#> 74    0.05178971
#> 75    0.05610830
#> 76    0.05525835
#> 77    0.05791255
#> 78    0.06829305
#> 79    0.03968171
#> 80    0.04508101
#> 81    0.04273005
#> 82    0.05702571
#> 83    0.04536797
#> 84    0.05707455
#> 85    0.05157984
#> 86    0.04071927
#> 87    0.05743992
#> 88    0.03888754
#> 89    0.04685289
#> 90    0.05578516
#> 91    0.04328590
#> 92    0.05473981
#> 93    0.06010586
#> 94    0.05848398
#> 95    0.04933253
#> 96    0.05063672
#> 97    0.05400358
#> 98    0.04947819
#> 99    0.04326903
#> 100   0.05187547
#> 101   0.04369832
#> 102   0.05031185
#> 103   0.05513101
#> 104   0.04594529
#> 105   0.04155457
#> 106   0.03834044
#> 107   0.05281728
#> 108   0.06132299
#> 109   0.04166210
#> 110   0.05284964
#> 111   0.05727677
#> 112   0.05125042
#> 113   0.06015421
#> 114   0.04109064
#> 115   0.06359701
#> 116   0.03894134
#> 117   0.05403589
#> 118   0.04227703
#> 119   0.04654525
#> 120   0.04507256
#> 121   0.04579421
#> 122   0.04894837
#> 123   0.04247355
#> 124   0.05518757
#> 125   0.04309374
#> 126   0.04142430
#> 127   0.04729118
#> 128   0.05013980
#> 129   0.06413022
#> 130   0.05190141
#> 131   0.05806102
#> 132   0.04864215
#> 133   0.05732324
#> 134   0.05157704
#> 135   0.06156832
#> 136   0.05605644
#> 137   0.05186794
#> 138   0.05267301
#> 139   0.06110643
#> 140   0.04464343
#> 141   0.04106200
#> 142   0.04859410
#>  [ reached getOption("max.print") -- omitted 9858 rows ]